+++ packages/article-skeleton/src/head.web.js
@@ -104,7 +104,28 @@ const getThumbnailUrlFromImage = article => {
   return get(article.leadAsset, "crop169.url", null);
 };
 
-function Head({ article, logoUrl, paidContentClassName }) {
+const getThumbnailUrl = article => {
+  const { hasVideo, leadAsset } = article;
+  const thumbnailUrl = hasVideo
+    ? getVideoLeadAssetUrl(article)
+    : getThumbnailUrlFromImage(article);
+
+  if (thumbnailUrl) return thumbnailUrl;
+
+  if (!leadAsset) return null;
+
+  const { crop32, crop1251, crop11, crop45, crop23, crop2251 } =
+    leadAsset && leadAsset.posterImage ? leadAsset.posterImage : leadAsset;
+  const crop = crop32 || crop1251 || crop11 || crop45 || crop23 || crop2251;
+  return crop ? crop.url : "";
+};
+
+function Head({
+  article,
+  logoUrl,
+  paidContentClassName,
+  getFallbackThumbnailUrl169
+}) {
   const {
     descriptionMarkup,
     headline,
@@ -126,19 +147,18 @@ function Head({ article, logoUrl, paidContentClassName }) {
       ? renderTreeAsText({ children: descriptionMarkup })
       : null);
   const sectionname = getSectionName(article);
-  const leadassetUrl = appendToImageURL(
-    getArticleLeadAssetUrl(article),
-    "resize",
-    685
-  );
+  const thumbnailUrl =
+    getThumbnailUrl(article) ||
+    (getFallbackThumbnailUrl169 ? getFallbackThumbnailUrl169() : null);
+
+  const leadassetUrl =
+    appendToImageURL(getArticleLeadAssetUrl(article), "resize", 1200) ||
+    thumbnailUrl;
   const authors = getAuthorSchema(article);
   const caption = get(leadAsset, "caption", null);
   const title = headline || shortHeadline || "";
   const datePublished = new Date(publishedTime).toISOString();
   const dateModified = updatedTime || datePublished;
-  const thumbnailUrl = hasVideo
-    ? getVideoLeadAssetUrl(article)
-    : getThumbnailUrlFromImage(article);
 
   const defaultAuthorSchema = {
     "@type": "Organization",
@@ -193,7 +213,7 @@ function Head({ article, logoUrl, paidContentClassName }) {
         description:
           Array.isArray(descriptionMarkup) && descriptionMarkup.length
             ? renderTreeAsText({ children: descriptionMarkup })
-            : null,
+            : seoDescription || leadAsset.title || title,
         contentUrl: url
       }
     : null;
@@ -208,6 +228,7 @@ function Head({ article, logoUrl, paidContentClassName }) {
               {title} | {sectionname ? `${sectionname} | ` : ""}
               {publication}
             </title>
+            <meta name="robots" content="max-image-preview:large" />
             <meta content={title} name="article:title" />
             <meta content={publication} name="article:publication" />
             {desc && <meta content={desc} name="description" />}
@@ -254,7 +275,8 @@ Head.propTypes = {
     tiles: PropTypes.array
   }).isRequired,
   logoUrl: PropTypes.string.isRequired,
-  paidContentClassName: PropTypes.string.isRequired
+  paidContentClassName: PropTypes.string.isRequired,
+  getFallbackThumbnailUrl169: PropTypes.func.isRequired
 };
 
 export default Head;