+++ packages/article-main-standard/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.35.0](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.34.40...@times-components/article-main-standard@3.35.0) (2021-08-23)
+
+
+### Features
+
+* **SCB-1193:** add aditional in article ad slots to article page logged in users ([#2758](https://github.com/newsuk/times-components/issues/2758)) ([481ff7f](https://github.com/newsuk/times-components/commit/481ff7f325b2bf0a216da583bf0d1c5871ef85f5))
+
+
+
+
+
+## [3.34.40](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.34.39...@times-components/article-main-standard@3.34.40) (2021-08-23)
+
+
+### Bug Fixes
+
+* **TSEO-31:** Updated image size on mobile to 1200px ([#2771](https://github.com/newsuk/times-components/issues/2771)) ([9a756ac](https://github.com/newsuk/times-components/commit/9a756acacf45c120de889b81823c591d890b4f38))
+
+
+
+
+
+## [3.34.39](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.34.38...@times-components/article-main-standard@3.34.39) (2021-08-23)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.34.38](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.34.37...@times-components/article-main-standard@3.34.38) (2021-08-23)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.34.37](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.34.36...@times-components/article-main-standard@3.34.37) (2021-08-20)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
+## [3.34.36](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.34.35...@times-components/article-main-standard@3.34.36) (2021-08-19)
+
+**Note:** Version bump only for package @times-components/article-main-standard
+
+
+
+
+
 ## [3.34.35](https://github.com/newsuk/times-components/compare/@times-components/article-main-standard@3.34.34...@times-components/article-main-standard@3.34.35) (2021-08-11)
 
 **Note:** Version bump only for package @times-components/article-main-standard