+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.52.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.51.43...@times-components/article-skeleton@1.52.0) (2021-08-23)
+
+
+### Features
+
+* **SCB-1193:** add aditional in article ad slots to article page logged in users ([#2758](https://github.com/newsuk/times-components/issues/2758)) ([481ff7f](https://github.com/newsuk/times-components/commit/481ff7f325b2bf0a216da583bf0d1c5871ef85f5))
+
+
+
+
+
+## [1.51.43](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.51.42...@times-components/article-skeleton@1.51.43) (2021-08-23)
+
+
+### Bug Fixes
+
+* **TSEO-31:** Updated image size on mobile to 1200px ([#2771](https://github.com/newsuk/times-components/issues/2771)) ([9a756ac](https://github.com/newsuk/times-components/commit/9a756acacf45c120de889b81823c591d890b4f38))
+
+
+
+
+
+## [1.51.42](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.51.41...@times-components/article-skeleton@1.51.42) (2021-08-23)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.51.41](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.51.40...@times-components/article-skeleton@1.51.41) (2021-08-23)
+
+
+### Bug Fixes
+
+* **TSEO-27:** updated video schema description ([#2770](https://github.com/newsuk/times-components/issues/2770)) ([64c5706](https://github.com/newsuk/times-components/commit/64c5706aa48e17a7a787400e0d4a075bcbbb7d83))
+
+
+
+
+
+## [1.51.40](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.51.39...@times-components/article-skeleton@1.51.40) (2021-08-20)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.51.39](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.51.38...@times-components/article-skeleton@1.51.39) (2021-08-19)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 ## [1.51.38](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.51.37...@times-components/article-skeleton@1.51.38) (2021-08-11)
 
 **Note:** Version bump only for package @times-components/article-skeleton